fix: Cast pooled sentence vectors to float32 before numpy conversion
- Updated the mean pooling in `_encode_hf_meanpool` to cast pooled sentence vectors to float32 to ensure compatibility with numpy.
- Modified `build_meaning_matrix_hf` to include a float cast for the mean vectors, addressing potential dtype issues with bfloat16 encoders.
- Enhanced `make_dataloader` to use a common dictionary for DataLoader parameters, improving readability and maintainability.

+++ src/zte/data/meaning.py
@@ -195,8 +195,10 @@ def build_meaning_matrix_hf(
         for sub, wp in enumerate(enc_in.word_ids(0)):
             if wp is not None and wp < len(positions):
                 pooled[wp].append(hs[sub])
+        # .float(): a bf16 encoder (some run in bfloat16 on GPU) has no numpy dtype; cast before numpy.
         vec_by_pos = {
-            positions[wp]: torch.stack(vecs).mean(0).cpu().numpy() for wp, vecs in pooled.items()
+            positions[wp]: torch.stack(vecs).mean(0).float().cpu().numpy()
+            for wp, vecs in pooled.items()
         }
         for i in rows:
             v = vec_by_pos.get(int(widx[i]))
